# openapi.yaml
openapi: 3.0.0
paths:
...
  /users:
    get:
      description: Retrieve a list of all users. Optionally filter by type using query parameter.
      operationId: UsersController_findall
      parameters:
        - name: type
          required: false
          in: query
          description: Filter users by type
          schema:
            enum:
              - user
              - admin
            type: string
      responses:
        "200":
          description: List of users retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserResponseDto"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "403":
          description: Forbidden (insufficient permissions)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "404":
          description: No users found with the specified type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
      security:
        - session-id: []
      summary: Get all users
      tags: &a2
        - users
    post:
      description: Create a new user with name, username, email, password, and type
      operationId: UsersController_create
      parameters: []
      requestBody:
        required: true
        description: User data to create
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserDto"
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponseDto"
        "400":
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "403":
          description: Forbidden (insufficient permissions)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "409":
          description: Conflict (e.g., duplicate entry)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "503":
          description: Service unavailable (e.g., database connection error)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
      security:
        - session-id: []
      summary: Create a new user
      tags: *a2
  /users/{id}:
    get:
      description: Retrieve a specific user by their unique identifier
      operationId: UsersController_findOne
      parameters:
        - name: id
          required: true
          in: path
          description: User ID (GUID)
          schema:
            example: 550e8400-e29b-41d4-a716-446655440000
            type: string
      responses:
        "200":
          description: User retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponseDto"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "403":
          description: Forbidden (insufficient permissions)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
      security:
        - session-id: []
      summary: Get user by ID
      tags: *a2
    patch:
      description: Update an existing user. Only provided fields will be updated.
      operationId: UsersController_update
      parameters:
        - name: id
          required: true
          in: path
          description: User ID (GUID)
          schema:
            example: 550e8400-e29b-41d4-a716-446655440000
            type: string
      requestBody:
        required: true
        description: User data to update (partial)
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserDto"
      responses:
        "200":
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponseDto"
        "400":
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "403":
          description: Forbidden (insufficient permissions)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "409":
          description: Conflict (e.g., duplicate entry)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "503":
          description: Service unavailable (e.g., database connection error)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
      security:
        - session-id: []
      summary: Update user
      tags: *a2
    delete:
      description: Delete a user by their unique identifier
      operationId: UsersController_remove
      parameters:
        - name: id
          required: true
          in: path
          description: User ID (GUID)
          schema:
            example: 550e8400-e29b-41d4-a716-446655440000
            type: string
      responses:
        "200":
          description: User deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponseDto"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "403":
          description: Forbidden (insufficient permissions)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "503":
          description: Service unavailable (e.g., database connection error)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
      security:
        - session-id: []
      summary: Delete user
      tags: *a2
...
info:
  title: Nest API Demo
  description: A comprehensive REST API built with NestJS and MySQL. This API provides endpoints for managing users, employees, and departments with role-based access control.
  version: "1.0"
  contact: {}
...
servers: []
components:
  securitySchemes:
    cookie:
      type: apiKey
      in: cookie
      name: session-id
  schemas:
...
    ErrorResponseDto:
      type: object
      properties:
        statusCode:
          type: number
          description: HTTP status code of the error response
          example: 401
        timestamp:
          type: string
          description: ISO 8601 timestamp of when the error occurred
          example: 2024-01-15T10:30:00.000Z
          format: date-time
        path:
          type: string
          description: Request path that produced the error
          example: /api/employees/123
        response:
          description: Error response payload (string message or an object returned by Nest exceptions)
          oneOf:
            - type: string
              example: Unauthorized
            - type: object
              additionalProperties: true
              example:
                message: Unauthorized
                statusCode: 401
      required:
        - statusCode
        - timestamp
        - path
        - response
    UserResponseDto:
      type: object
      properties:
        id:
          type: string
          description: The unique identifier of the user (GUID)
          example: 550e8400-e29b-41d4-a716-446655440000
        name:
          type: string
          description: The full name of the user
          example: John Doe
        username:
          type: string
          description: The username of the user
          example: johndoe
        email:
          type: string
          description: The email address of the user
          example: john.doe@example.com
          format: email
        type:
          type: string
          description: The type of the user
          enum:
            - user
            - admin
          example: user
        createdAt:
          type: string
          description: The date and time when the user was created
          example: 2024-01-15T10:30:00.000Z
          format: date-time
        updatedAt:
          type: string
          description: The date and time when the user was last updated
          example: 2024-01-15T10:30:00.000Z
          format: date-time
      required:
        - id
        - name
        - username
        - email
        - type
        - createdAt
        - updatedAt
    CreateUserDto:
      type: object
      properties:
        name:
          type: string
          description: The full name of the user
          example: John Doe
          minLength: 1
        username:
          type: string
          description: The username of the user
          example: johndoe
          minLength: 3
        email:
          type: string
          description: The email address of the user
          example: john.doe@example.com
          format: email
        password:
          type: string
          description: The password of the user
          example: SecurePassword123!
          minLength: 6
        type:
          type: string
          description: The type of the user
          enum: &a6
            - user
            - admin
          example: user
      required:
        - name
        - username
        - email
        - password
        - type
    UpdateUserDto:
      type: object
      properties:
        name:
          type: string
          description: The full name of the user
          example: John Doe
          minLength: 1
        username:
          type: string
          description: The username of the user
          example: johndoe
          minLength: 3
        email:
          type: string
          description: The email address of the user
          example: john.doe@example.com
          format: email
        password:
          type: string
          description: The password of the user
          example: SecurePassword123!
          minLength: 6
        type:
          type: string
          description: The type of the user
          enum: *a6
          example: user
...